pyproject.toml
[project]
name = "sic_core"
version = "0.1.0"
description = "Implementation of accounting system based in hordak and following SIC (Andrisani) definitions."
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "gastong256" }]

dependencies = [
    "django>=5.1,<6",
    "djangorestframework>=3.15",
    "drf-spectacular>=0.27",
    "django-environ>=0.11",
    "django-filter>=24",
    "psycopg[binary]>=3.1",
    "gunicorn>=22",
    "xlsxwriter>=3.2",
    "structlog>=24",
    "whitenoise>=6",
    "django-hordak>=2.0.0",
    "djangorestframework-simplejwt>=5.5.1",
    "django-cors-headers>=4.9.0",
]

[project.optional-dependencies]
otel = [
    "opentelemetry-sdk>=1.25",
    "opentelemetry-instrumentation-django>=0.48b0",
    "opentelemetry-instrumentation-psycopg>=0.48b0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.25",
]

[dependency-groups]
dev = [
    "pytest>=8",
    "pytest-django>=4.9",
    "pytest-cov>=5",
    "factory-boy>=3.3",
    "ruff>=0.6",
    "black>=24",
    "pyright>=1.1",
    "pre-commit>=3",
    "openpyxl>=3.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps", "config"]

[tool.ruff]
target-version = "py312"
line-length = 100
extend-exclude = [
    "docs/openapi",
    "*/migrations/*",
]

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["B008", "N806", "E501"]

[tool.ruff.lint.isort]
known-first-party = ["apps", "config"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings.test"
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--strict-markers -v --tb=short"
testpaths = ["tests"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "off"
include = ["apps", "config", "tests"]
exclude = [
    "**/migrations/**",
    "tests/**",
    "apps/**/api/**",
    "apps/**/admin.py",
]
venvPath = "."
venv = ".venv"

[tool.coverage.run]
source = ["apps", "config"]
omit = ["**/migrations/**", "**/tests/**"]

[tool.coverage.report]
skip_empty = true
show_missing = true

[tool.semantic_release]
version_toml = ["pyproject.toml:project.version"]
branch = "main"
changelog_file = "CHANGELOG.md"
build_command = ""
upload_to_pypi = false
upload_to_release = true
tag_format = "v{version}"
commit_message = "chore(release): v{version} [skip ci]"
